source/bot/requirements.txt
prometheus-client
aiogram>=3.0
python-dotenv
PyYAML
python-logging-loki
